requirements.txt
# Product Catalog Dependencies
Flask==2.3.3
Pillow==10.0.1
XlsxWriter==3.1.9
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
